requirements.txt
Appium-Python-Client==2.11.1
pytest-assume==2.4.3
attrs==23.1.0
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
qase-pytest==5.0.2
qase-python-commons==2.0.3
qaseio==3.4.1
requests==2.31.0
selenium==4.11.2